refactor: make AIToolService override method chatMessageDicts()
