+++ Easydict/Swift/Service/AITool/AIToolService.swift
@@ -10,9 +10,7 @@ import Foundation
 
 /// A class used for LLM derivatives such as summary and polishing
 /// Based on `BuiltInAIService` and takes `llama3-70b-8192` as the LLM
-class LLMDerivService: BuiltInAIService {
-    // MARK: Public
-
+class AIToolService: BuiltInAIService {
     public override func configurationListItems() -> Any {
         StreamConfigurationView(
             service: self,
@@ -26,25 +24,5 @@ class LLMDerivService: BuiltInAIService {
             showDictionaryToggle: false,
             showUsageStatusPicker: true
         )
-    }
-
-    // MARK: Internal
-
-    override var defaultModels: [String] {
-        ["llama3-70b-8192"]
-    }
-
-    func serviceChatMessage(_ chatQuery: LLMDerivParam) -> [Any] {
-        var chatModels: [ChatMessage] = []
-        for message in llmDerivMessageDicts(chatQuery) {
-            if let roleRawValue = message["role"],
-               let role = ChatMessage.Role(rawValue: roleRawValue),
-               let content = message["content"] {
-                if let chat = ChatMessage(role: role, content: content) {
-                    chatModels.append(chat)
-                }
-            }
-        }
-        return chatModels
     }
 }
